Refactor authentication and user validation***
This commit refactors the authentication logic in the `authControllers.js` file. It replaces the check for a null user with a more explicit comparison. Additionally, the commit updates the JWT secret to use the `APP_SECRET` environment variable.

In the `userMiddleware.js` file, the commit simplifies the user validation schema by removing unnecessary properties and constraints.